+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - add support for single group plot
 - add shiny application for mass defect analysis
+- add function to compute high order mass defect
 
 # enviGCMS 0.5.0