+++ NEWS.md
@@ -6,6 +6,8 @@
 - add shiny application for mass defect analysis
 - add function to compute high order mass defect
 - add function to compare two peak list by overlap
+- add function to output csv file from list object
+- fix wrong object name in getmzrt/getmzrt2
 
 # enviGCMS 0.5.0